+++ libsidx/datastream.hpp
@@ -30,13 +30,13 @@
 class DataStream : public SpatialIndex::IDataStream
 {
 public:
-    DataStream(int (*readNext)(SpatialIndex::id_type* id, double **pMin, double **pMax, uint32_t *nDimension, const uint8_t **pData, size_t *nDataLength));
+    DataStream(int (*readNext)(SpatialIndex::id_type* id, double **pMin, double **pMax, uint32_t *nDimension, const uint8_t **pData, uint32_t *nDataLength));
     ~DataStream();
 
     SpatialIndex::IData* getNext();
     bool hasNext() throw (Tools::NotSupportedException);
 
-    size_t size() throw (Tools::NotSupportedException);
+    uint32_t size() throw (Tools::NotSupportedException);
     void rewind() throw (Tools::NotSupportedException);
 
 protected:
@@ -44,7 +44,7 @@ protected:
     SpatialIndex::id_type m_id;
 
 private:
-    int (*iterfunct)(SpatialIndex::id_type *id, double **pMin, double **pMax, uint32_t *nDimension, const uint8_t **pData, size_t *nDataLength);
+    int (*iterfunct)(SpatialIndex::id_type *id, double **pMin, double **pMax, uint32_t *nDimension, const uint8_t **pData, uint32_t *nDataLength);
     
     bool readData();
     bool m_bDoneReading;
